feat: refactor voting logic; move nominee merging and finalization to AdminVoteService
